feat: stafe on jump feature
